# swagger/swagger.yaml
consumes:
    - application/json
definitions:
...
    DeletedAt:
        $ref: '#/definitions/NullTime'
    DeployInstanceRequest:
        properties:
            groupName:
                type: string
                x-go-name: Group
            name:
                type: string
                x-go-name: Name
            optionalParameters:
                items:
                    $ref: '#/definitions/InstanceOptionalParameter'
                type: array
                x-go-name: OptionalParameters
            presetInstance:
                format: uint64
                type: integer
                x-go-name: PresetInstance
            requiredParameters:
                items:
                    $ref: '#/definitions/InstanceRequiredParameter'
                type: array
                x-go-name: RequiredParameters
            sourceInstance:
                format: uint64
                type: integer
                x-go-name: SourceInstance
            stackName:
                type: string
                x-go-name: Stack
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/instance
...
    GroupWithInstances:
        properties:
            Hostname:
                type: string
            Instances:
                items:
                    $ref: '#/definitions/Instance'
                type: array
            Name:
                type: string
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/instance
...
    Instance:
        properties:
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            DeployLog:
                type: string
            GroupName:
                type: string
            ID:
                format: uint64
                type: integer
            Name:
                type: string
            Preset:
                type: boolean
            PresetID:
                format: uint64
                type: integer
            StackName:
                type: string
            UpdatedAt:
                format: date-time
                type: string
            UserID:
                format: uint64
                type: integer
            optionalParameters:
                items:
                    $ref: '#/definitions/InstanceOptionalParameter'
                type: array
                x-go-name: OptionalParameters
            requiredParameters:
                items:
                    $ref: '#/definitions/InstanceRequiredParameter'
                type: array
                x-go-name: RequiredParameters
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
    InstanceOptionalParameter:
        properties:
            name:
                description: 'TODO: Rename StackOptionalParameterID to Name'
                type: string
                x-go-name: StackOptionalParameterID
            value:
                type: string
                x-go-name: Value
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
    InstanceRequiredParameter:
        properties:
            name:
                description: 'TODO: Rename StackRequiredParameterID to Name'
                type: string
                x-go-name: StackRequiredParameterID
            value:
                type: string
                x-go-name: Value
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
    NullTime:
        description: |-
            NullTime implements the Scanner interface so
            it can be used as a scan destination, similar to NullString.
        properties:
            Time:
                format: date-time
                type: string
            Valid:
                type: boolean
        title: NullTime represents a time.Time that may be null.
        type: object
        x-go-package: database/sql
...
    UpdateInstanceRequest:
        properties:
            optionalParameters:
                items:
                    $ref: '#/definitions/InstanceOptionalParameter'
                type: array
                x-go-name: OptionalParameters
            requiredParameters:
                items:
                    $ref: '#/definitions/InstanceRequiredParameter'
                type: array
                x-go-name: RequiredParameters
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/instance
...
info:
    contact:
        email: info@dhis2.org
        url: https://github.com/dhis2-sre/im-manager
    description: Instance Manager
    license:
        name: TODO
    title: Instance Manager
    version: 0.1.0
paths:
...
    /instances:
        get:
            description: List all instances accessible by the user
            operationId: listInstances
            responses:
                "200":
                    $ref: '#/responses/GroupWithInstances'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: List instances
        post:
            description: Deploy an instance...
            operationId: deployInstance
            parameters:
                - description: Deploy instance request body parameter
                  in: body
                  name: Payload
                  required: true
                  schema:
                    $ref: '#/definitions/DeployInstanceRequest'
                - description: preset
                  in: query
                  name: preset
                  type: string
                  x-go-name: Preset
            responses:
                "201":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Deploy instance
    /instances-name-to-id/{groupName}/{instanceName}:
        get:
            description: Find instance id by name and group name
            operationId: instanceNameToId
            parameters:
                - in: path
                  name: groupName
                  required: true
                  type: string
                  x-go-name: GroupName
                - in: path
                  name: instanceName
                  required: true
                  type: string
                  x-go-name: InstanceName
            responses:
                "200":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Find an instance
    /instances/{id}:
        delete:
            description: Delete an instance by id
            operationId: deleteInstance
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "202":
                    description: ""
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Delete instance
        get:
            description: Find an instance by id
            operationId: findById
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Find instance
        put:
            description: Update an instance...
            operationId: updateInstance
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
                - description: Update instance request body parameter
                  in: body
                  name: Payload
                  required: true
                  schema:
                    $ref: '#/definitions/UpdateInstanceRequest'
            responses:
                "204":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Update instance
    /instances/{id}/logs:
        get:
            description: Stream instance logs in real time
            operationId: instanceLogs
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
                - description: selector
                  in: query
                  name: selector
                  type: string
                  x-go-name: Selector
            responses:
                "200":
                    $ref: '#/responses/InstanceLogsResponse'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Stream logs
    /instances/{id}/parameters:
        get:
            description: Find instance by id with decrypted parameters
            operationId: findByIdDecrypted
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Find decrypted instance
    /instances/{id}/pause:
        put:
            description: |-
                Pause an instance. Pause can be called multiple times even on an already paused instance
                (idempotent).
            operationId: pauseInstance
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "202":
                    description: ""
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Pause instance
    /instances/{id}/restart:
        put:
            description: Restart an instance...
            operationId: restartInstance
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
                - description: selector
                  in: query
                  name: selector
                  type: string
                  x-go-name: Selector
            responses:
                "202":
                    description: ""
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Restart instance
    /instances/{id}/resume:
        put:
            description: |-
                Resume a paused instance. Resume can be called multiple times even on an already running
                instance (idempotent).
            operationId: resumeInstance
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "202":
                    description: ""
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Resume paused instance
...
    /presets:
        get:
            description: List all presets accessible by the user
            operationId: listPresets
            responses:
                "200":
                    $ref: '#/responses/GroupWithInstances'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: List presets
...
produces:
    - application/json
responses:
...
    Error:
        description: ""
    GroupWithInstances:
        description: ""
        schema:
            items:
                $ref: '#/definitions/GroupWithInstances'
            type: array
    InstanceLogsResponse:
        description: ""
...
securityDefinitions:
    oauth2:
        flow: password
        tokenUrl: /tokens
        type: oauth2
swagger: "2.0"